refactor: create terraform package
