+++ get.go
@@ -5,6 +5,7 @@ import (
 	"log"
 
 	"github.com/tmknom/tfmod/internal/format"
+	"github.com/tmknom/tfmod/internal/terraform"
 )
 
 type GetRunner struct {
@@ -46,8 +47,8 @@ func (r *GetRunner) Run() error {
 func (r *GetRunner) TerraformGet() ([]string, error) {
 	log.Printf("Runner flags: %#v", r.flags)
 
-	terraform := NewTerraform(r.flags.GetBaseDir(), r.flags.EnableTf)
-	sourceDirs, err := terraform.GetAll()
+	terraformCommand := terraform.NewTerraform(r.flags.GetBaseDir(), r.flags.EnableTf)
+	sourceDirs, err := terraformCommand.GetAll()
 	if err != nil {
 		return nil, err
 	}
